add a trait `conn::Reader` / `conn::Writer`
